[parser] remove namespace
